surface: make generic against memory descriptor type
Make Surface take ownership of a memory descriptor that represents its
backing memory. This allows us to easily implement different types of
memory backing by having them implement SurfaceMemoryDescriptor.

() is the type used to let VA allocate and manage the memory. In the
future, we will add implementations for AsRef<[u8]> (for user-backed
memory) and DMABufs.

SurfaceMemoryDescriptor could also be implemented on an enum type that
is a composite of other implementors of SurfaceMemoryDescriptor in order
to allow the user to choose the memory type at runtime. It should even
be implementable on trait objects.